feat: native intersection observer reveals and view transitions
- Sync branch with origin/main to preserve all merged PR refactoring.
- Audit portfolio for scroll animations and clean up legacy scroll listeners.
- Replace manual scroll listener in ScrollProgress with declarative CSS scroll timeline, ensuring un-supported browsers get a full-width static progress indicator.
- Implement reusable ScrollReveal component leveraging native Intersection Observer with feature detection and batch index-based stagger.
- Integrate ScrollReveal and lazy GSAP triggers for About, Project, and Blog lists.
- Set up native Next.js View Transitions between Project list and Project Detail page.
- Clean up duplicate hero rendering in media galleries.
- Ensure perfect fallback rendering and prefers-reduced-motion support.
