+++ package.json
@@ -52,12 +52,12 @@
     "shiki": "^4.3.1"
   },
   "devDependencies": {
-    "@biomejs/biome": "^2.5.3",
+    "@biomejs/biome": "^2.5.4",
     "@playwright/test": "^1.61.1",
-    "@tailwindcss/postcss": "^4.3.2",
+    "@tailwindcss/postcss": "^4.3.3",
     "husky": "^9.1.7",
     "lint-staged": "^17.0.8",
-    "tailwindcss": "^4.3.2"
+    "tailwindcss": "^4.3.3"
   },
   "packageManager": "pnpm@10.0.0",
   "lint-staged": {
