--- server.log
@@ -1,8 +1,0 @@
-
-> justinjdaniel.com@5.0.0 start /app
-> next start
-
-▲ Next.js 16.2.10
-- Local:         http://localhost:3000
-- Network:       http://192.168.0.2:3000
-✓ Ready in 183ms
