+++ src/components/mdx/pros-card.js
@@ -1,15 +1,7 @@
-/**
- * ProsCard Component
- * @description This component is used to display a card with a list of pros for a specific technology or topic.
- * @param {string} title - The title of the card.
- * @param {Array<string>} pros - An array of strings representing the pros.
- * @returns {JSX.Element} The ProsCard component.
- * @example
- * <ProsCard title="React" pros={["Fast", "Flexible", "Popular"]} />
- * // Renders a card with the title "React" and a list of pros.
- */
+// MARK: - Imports
 import React from "react";
 
+// MARK: - Config & Constants
 const checkIcon = (
   <svg
     className="w-5 h-5 text-emerald-500 flex-shrink-0 mt-0.5"
@@ -37,6 +29,15 @@ const checkIcon = (
   </svg>
 );
 
+// MARK: - Helper Functions
+
+/**
+ * Injects a React icon element recursively into a list structure.
+ *
+ * @param {import("react").ReactNode} children - Component nested child nodes.
+ * @param {import("react").ReactNode} icon - SVG/HTML icon payload to display on items.
+ * @returns {import("react").ReactNode} Modified virtual DOM node structure.
+ */
 function injectIcon(children, icon) {
   return React.Children.map(children, (child) => {
     if (!child || !React.isValidElement(child)) return child;
@@ -70,6 +71,18 @@ function injectIcon(children, icon) {
   });
 }
 
+// MARK: - Render
+
+/**
+ * ProsCard Component - displays a summary list of positive features or topics.
+ *
+ * @component
+ * @param {Object} props - Component property arguments.
+ * @param {string} [props.title] - Name of subject/technology.
+ * @param {Array<string>} [props.pros=[]] - List items.
+ * @param {import("react").ReactNode} [props.children] - Children list fallback.
+ * @returns {import("react").JSX.Element}
+ */
 export default function ProsCard({ title, pros = [], children }) {
   return (
     <div
