+++ src/lib/db/queries/blog-post-views.js
@@ -1,21 +1,38 @@
+// MARK: - Imports
 import { ensureTable, pool } from "../db";
 
+// MARK: - Config & Constants
 const isMock = !pool;
 
-// Initialization promise to ensure all required tables exist before any DB access
-const dbInitPromise = ensureTable(
-  "blog_views",
-  "CREATE TABLE IF NOT EXISTS blog_views (slug TEXT PRIMARY KEY, count INTEGER DEFAULT 0);",
-);
+// Lazy initialization logic
+let dbInitPromise = null;
+
+/**
+ * Lazy helper to ensure all required tables exist before any DB access is made.
+ *
+ * @returns {Promise<void>}
+ */
+function initializeDatabase() {
+  if (!dbInitPromise) {
+    dbInitPromise = ensureTable(
+      "blog_views",
+      "CREATE TABLE IF NOT EXISTS blog_views (slug TEXT PRIMARY KEY, count INTEGER DEFAULT 0);",
+    );
+  }
+  return dbInitPromise;
+}
+
+// MARK: - Helper Functions
 
 /**
  * Increments the view count for a blog post and returns the updated count.
+ *
  * @param {string} slug - The blog post slug.
  * @returns {Promise<number|null>} The updated view count or null on error.
  */
 export async function incrementAndGetBlogViewCount(slug) {
   if (isMock) return 0;
-  await dbInitPromise;
+  await initializeDatabase();
   let client;
   try {
     client = await pool.connect();
@@ -41,15 +58,15 @@ export async function incrementAndGetBlogViewCount(slug) {
   }
 }
 
-// ----------- Sample functions for mutation and query ------- //
 /**
  * Increments the view count for a blog post (does not return the count).
+ *
  * @param {string} slug - The blog post slug.
  * @returns {Promise<void>}
  */
 export async function incrementBlogViewCount(slug) {
   if (isMock) return;
-  await dbInitPromise;
+  await initializeDatabase();
   let client;
   try {
     client = await pool.connect();
@@ -75,12 +92,13 @@ export async function incrementBlogViewCount(slug) {
 
 /**
  * Gets the view count for a blog post.
+ *
  * @param {string} slug - The blog post slug.
  * @returns {Promise<number|null>} The current view count or null if not found.
  */
 export async function getBlogViewCount(slug) {
   if (isMock) return 0;
-  await dbInitPromise;
+  await initializeDatabase();
   let client;
   try {
     client = await pool.connect();
